+++ fp-plugins/cookiebanner/lang/lang.es-es.php
@@ -1,0 +1,9 @@
+<?php
+$baseurl = BLOG_BASEURL;
+$lang ['plugin'] ['cookiebanner'] = array(
+	'bannertext' => 'Utilizo <a href="https://es.wikipedia.org/wiki/Cookie_(inform%C3%A1tica)" title="Wikipedia" target="_blank">cookies</a> ' . //
+		'para garantizar que este sitio web funcione de forma fiable y segura.',
+	'ok' => 'ACEPTO',
+	'notice_text' => 'Trato sus datos de acuerdo con mi <a href="' . $baseurl . 'static.php?page=privacy-policy">política de privacidad</a>.'
+);
+?>